moved application code out of engine dirs'
