--- app/controllers/places/application_controller.rb
@@ -1,4 +1,0 @@
-module Places
-  class ApplicationController < ActionController::Base
-  end
-end
