--- app/helpers/places/places_helper.rb
@@ -1,11 +1,0 @@
-module Places
-    module PlacesHelper
-
-        def recurs_places(nodes)
-            nodes.each do |n|
-                render partial: 'place', locals: { node: n }
-            end
-        end
-
-    end
-end
