--- app/views/layouts/places/application.html.erb
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Places</title>
-  <%= stylesheet_link_tag    "places/application", media: "all" %>
-  <%= javascript_include_tag "places/application" %>
-  <%= csrf_meta_tags %>
-</head>
-<body>
-
-<%= display_flash_messages %> 
-<%= display_flash_errors %>
-
-<%= yield %>
-
-</body>
-</html>
