+++ spec/dummy/app/assets/stylesheets/application.css
@@ -13,3 +13,7 @@
  *= require_tree .
  *= require_self
  */
+
+ .nested_place {
+    margin-left: 30px;
+ }
