+++ spec/dummy/app/helpers/places_helper.rb
@@ -1,0 +1,9 @@
+module PlacesHelper
+
+    def recurs_places(nodes)
+        nodes.each do |n|
+            render partial: 'place', locals: { node: n }
+        end
+    end
+
+end
