+++ spec/dummy/app/views/layouts/application.html.erb
@@ -8,6 +8,9 @@
 </head>
 <body>
 
+<%= display_flash_messages %> 
+<%= display_flash_errors %>
+
 <%= yield %>
 
 </body>
