+++ spec/dummy/app/views/places/_place.html.erb
