+++ spec/dummy/app/views/places/edit.html.erb
@@ -47,5 +47,5 @@
 
 <p><ul>
 <li><%= link_to "Select a Place", places_path %></li>
-<li><%= link_to "Create a new Place", create_path %></li>
+<li><%= link_to "Create a new Place", new_place_path %></li>
 </ul></p>
