+++ spec/dummy/app/views/places/index.html.erb
@@ -11,5 +11,5 @@
 <br/>
 <p><ul>
 <li><%= link_to "Select a Place", places_path %></li>
-<li><%= link_to "Create a new Place", create_path %></li>
+<li><%= link_to "Create a new Place", new_place_path %></li>
 </ul></p>
