+++ spec/dummy/app/views/places/new.html.erb
@@ -12,5 +12,5 @@
         <%= f.text_field :description %>
     </p>
   
-    <%= f.submit "Create a place!" %>
+    <%= f.submit %>
 <% end %>
