+++ spec/dummy/app/views/places/show.html.erb
@@ -22,7 +22,7 @@
 
 <p><ul>
 <li><%= link_to "Select a Place", places_path %></li>
-<li><%= link_to "Create a new Place", create_path %></li>
+<li><%= link_to "Create a new Place", new_place_path %></li>
 <li><%= link_to "Edit this place", edit_place_path(@place.id) %></li>
 
 </ul></p>
